exercicio 50 refeito codigo - treino
